Remove panic-expecting tests again
Because unfortunately neither ValidatePtr(track) nor
ValidatePtr2(null, track) return true when checking a
MediaTrack from another project tab. ValidatePtr2 would
return true if we would pass it the track's project. But
in general we don't have that in the medium-level API.
So on this level there's no way to validate MediaTrack pointers
and therefore also no way of failing fast if the pointer
is invalid. We need to mark those methods as unsafe.