Throw away most of today's code
It started with a good idea but it turned into something
unnecessarily complex. The idea was to make get_set_media_track_info()
itself more typesafe by combining the key and the value into one
thing. Similar to validate_ptr_2(). However, enums didn't cut it
because unlike validate_ptr_2() we have a result type which relates
to the input type. So we must use structs. But they are also not enough
because the different track info types need very different kinds of
conversions. With a combination of traits and structs we would have
been able to make it work perfectly. But this solution is not
straight-forward anymore. It's much simpler to implement the old
approach of having different `ReaperFunction` methods for each
track info key. And it's also easier to use because one can use
plain old auto completion and help to have every option immediately
visible.