High-level API: Don't distinguish between ReaperSession and Reaper
This distinction makes sense for the medium-level API because it leaves it
to the API consumer if globals are used or not. The high-level API goes global
all the way. It doesn't work without it. So I don't see why anyone would need
a separate struct which is easily cloneable. Result: The API get's more
straight-forward because there's only one important `get()` function = one static
variable.

We still gained something from that refactoring: It's now more clear which
functions need the advanced RefCell fields and which don't. At the moment, the
ones which don't, are implemented in `reaper_simple.rs`. The ones which do,
are implemented in `reaper.rs` and have been equipped with strict thread checks
where necessary.