Hopefully fix segmentation fault when running integration test on macOS
There's a description of what happened in `ReaperSession::plugin_register_remove_csurf_inst().`

Summary: The helper control surface unregistered itself within the `run()` method as part of
a test step where `reaper_high::Reaper::deactivate()` was called (now `go_to_sleep()`).
The problem was that the medium-level API not just unregistered the control surface but also
dropped it - which was too early because the `run()` method was not finished yet iterating
over tasks from a channel. Doing that when the channel receiver value is already dropped is
of course deadly.

The implemented solution in the medium-level API is to not drop the control surface but to give
ownership back to the consumer. The solution in the high-level API is to happily gain ownership
again and store the control surface until someone calls `wake_up()` and it will be registered
again. This design of swapping ownership also seems much better and logical.