Medium-level API: Equip BorrowedPcmSource with reference instead of pointer
Mainly in order to be able to express lifetime relations without having to
return a reference. Although we don't need that now it might very well be
necessary in future. With the MidiEventList type it's necessary already now
and it's probably better to use a similar approach everywhere. Last but not
least it requires less unsafe to create the reference only once at construction
time instead of in each BorrowedPcmSource method separately